Retrieve Assays from PubChem

This function sends a request to PubChem to retrieve assay data based on the specified parameters. It returns a list of assays corresponding to the provided identifiers.

Usage

get_assays(
  identifier,
  namespace = "aid",
  operation = NULL,
  searchtype = NULL,
  options = NULL
)

Arguments

identifier

A vector of positive integers (e.g. cid, sid, aid) or identifier strings (source, inchikey, formula). In some cases, only a single identifier string (name, smiles, xref; inchi, sdf by POST only).

namespace

Specifies the namespace for the query. For the 'compound' domain, possible values include 'cid', 'name', 'smiles', 'inchi', 'sdf', 'inchikey', 'formula', 'substructure', 'superstructure', 'similarity', 'identity', 'xref', 'listkey', 'fastidentity', 'fastsimilarity_2d', 'fastsimilarity_3d', 'fastsubstructure', 'fastsuperstructure', and 'fastformula'. For other domains, the possible namespaces are domain-specific.

operation

The operation to be performed (default: NULL).

searchtype

The type of search to be performed (default: NULL).

options

Additional parameters. Currently has no effect on the results.

Value

A named list where each element corresponds to an assay retrieved from PubChem. The names of the list elements are based on the provided identifiers. If no assay is found for a given identifier, the corresponding list element will contain the string "No assay".

Examples

get_assays(
  identifier = 1234,
  namespace = "aid"
)
#> $`'1234'`
#> $`'1234'`[[1]]
#> $`'1234'`[[1]]$assay
#> $`'1234'`[[1]]$assay$descr
#> $`'1234'`[[1]]$assay$descr$aid
#>      id version 
#>    1234       1 
#> 
#> $`'1234'`[[1]]$assay$descr$aid_source
#> $`'1234'`[[1]]$assay$descr$aid_source$db
#> $`'1234'`[[1]]$assay$descr$aid_source$db$name
#> [1] "University of Pittsburgh Molecular Library Screening Center"
#> 
#> $`'1234'`[[1]]$assay$descr$aid_source$db$source_id
#>                                                                               str 
#> "NIH Compound Library Profiling: Redox Cycling H2O2 Generation EC50 Confirmation" 
#> 
#> 
#> 
#> $`'1234'`[[1]]$assay$descr$name
#> [1] "NIH Compound Library Profiling: Compound and DTT Dependent Redox Cycling H2O2 Generation 10-point Concentration Response Confirmation Assay."
#> 
#> $`'1234'`[[1]]$assay$descr$description
#> [1] "List of compounds to be tested: active compounds identified in the Redox Cycling H2O2 Generation primary screen AID 878, will be confirmed in 10-point concentration response assays."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    
#> [2] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [3] "Hydrogen peroxide (H2O2) can modulate (activate or inhibit) the activity of a variety of proteins including protein kinases, protein phosphatases, transcription factors, phospholipases, ion channels and G proteins.  H2O2 is capable of oxidizing the cysteine residues of proteins that may be crucial for their catalytic and/or structural function. Many proteins, including a significant number of the targets screened by the MLSCN, contain an active site cysteine that is required for biological activity e.g. the protein tyrosine phosphatases, the cathepsin lysosomal proteases, the caspases (cysteine using, aspartic acid selective proteases), the ubiquitin-conjugating enzymes, and the peroxiredoxins. In biochemical assays for enzymes with active site cysteines, it is common practice to try to maintain the protein in its reduced and active form by including one of a variety of reducing agents in assay and/or enzyme storage buffers; dithiotreitol (DTT), tris(2-carboxyethyl)phosphine (TCEP), -mercaptoethanol (BME), glutathione (GSH) or cysteine (Cys). However, in the presence of oxygen at neutral to acidic pH DTT undergoes a chain oxidation reaction to produce H2O2, and furthermore a number of compounds have been demonstrated to generate H2O2 by a cyclic redox mechanism in the presence of reducing agents. "                                                                                                                                                                                                                                                                                                                                                                                                                                   
#> [4] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [5] "H2O2 is a relatively mild oxidant that can oxidize the cysteine residues of proteins to cysteine sulfenic acid (Cys-SOH) or disulfide that can reduced back to cysteine by cellular thiol donors such as GSH, Grx and Trx . The pKa of sulfhydryl groups of most cysteine residues (Cys-SH) in proteins is ~ 8.5 and they do not react at physiologically significant rates with H2O2 unless the reaction is catalyzed. The cysteine thiolate anion (Cys-S-) is much more readily oxidized by H2O2 and for some signaling proteins, most notably the protein tyrosine phsophatases (PTPs), the cysteine has been clearly shown to be preferentially in a thiolate form. The 107 protein tyrosine phosphatases (PTPs) found in the human genome are defined by the active site sequence C(X)5R(S/T), with X being any amino acid, and they are critical regulators of mammalian cell proliferation, differentiation, and apoptosis. The active site cysteine of PTPs is required for catalytic activity and performs a nucleophilic attack on the phosphotyrosine residues of the substrate to form a covalent thiol-phosphate intermediate followed by hydrolysis and the release of the phosphate.  Due to its unique environment and the invariant arginine, the pKa of PTP active site cysteines is unusually low (pKa 4.7-5.4) and exists as a thiolate anion (Cys-S-) at neutral pH, which enhances its catalytic function as a nucleophile but renders it susceptible to inactivation by H2O2. The specific and reversible oxidation and inactivation of PTPs by H2O2 has now been demonstrated for a number of PTPs involved in the regulation of key signaling pathways controlling normal physiological processes and disease progression; PTP1B, PTP, LAR, VHR, PTEN, SHP-2, Cdc25B, and MKP-3."
#> [6] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         
#> [7] "We have developed a simple 384-well homogeneous colorimetric assay to measure the generation H2O2 by compounds of the NIHs MLSCN small molecule library in the presence of DTT. It is anticipated that by profiling the full diversity MLSCN compound library in this assay we will identify compounds capable of generating H2O2 in the presence of reducing agents, and can annotate the PubChem database with this activity.  Through a simple cross target query of the PubChem database, these data will alert investigators whether their active compounds have the potential to modulate the activity of their targets through the redox cycling generation of H2O2. For target proteins with cysteine residues crucial for their catalytic and/or structural function this will identify false actives that modulate activity through an indirect mechanism and would not be worth pursuing."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#> 
#> $`'1234'`[[1]]$assay$descr$protocol
#>  [1] "Assay Protocol for H2O2 Generation Assay"                                                                                                                                                                             
#>  [2] "The method is based on the H2O2 and horseradish peroxidase (HRP) mediated oxidation of phenol red that alters its absorbance at 610 nm after the pH has been adjusted to alkaline pH with NaOH (Pick & Keisari 1981)."
#>  [3] ""                                                                                                                                                                                                                     
#>  [4] "Materials: "                                                                                                                                                                                                          
#>  [5] "1. H2O2 30 % (w/w) Sigma Aldrich CAS  7722-84-1 (8.82M)"                                                                                                                                                              
#>  [6] "2. Phenol Red Sigma Aldrich, catalog  P-2417. Prepare stock in 1.0 mg/mL in DI H2O."                                                                                                                                  
#>  [7] "3. Horseradish peroxidase,catalog  P2088-10KU. Prepare stock 2 mg/mL in Hanks Balanced Salt Solution (HBSS)"                                                                                                          
#>  [8] "4. Hanks Balanced Salt Solution (HBSS)."                                                                                                                                                                              
#>  [9] "5. DTT 200 mM stock. "                                                                                                                                                                                                
#> [10] "6. Stop reaction 1N NaOH."                                                                                                                                                                                            
#> [11] "7. Compounds in DMSO; typically 10 or 1 mM."                                                                                                                                                                          
#> [12] "8. 384-well 120 uL Greiner clear bottom plates."                                                                                                                                                                      
#> [13] ""                                                                                                                                                                                                                     
#> [14] "Protocol:"                                                                                                                                                                                                            
#> [15] "1. Prepare 3X Phenol Red-HRP detection reagent in HBSS. "                                                                                                                                                             
#> [16] "For 10 mL: HBSS6.1 mL + Phenol Red 3.0 mL (Stock 1.0 mg/mL)+ HRP 900 uL(Stock 2 mg/mL)"                                                                                                                               
#> [17] "2. Prepare 2.4 mM DTT in HBSS (3X concentration for 0.8 mM final DTT concentration in a 60 uL reaction volume)"                                                                                                       
#> [18] "For 10 mL: HBSS 9.88 mL + DTT 120 uL (stock 200 mM)"                                                                                                                                                                  
#> [19] "3. Dilute compounds in HBSS to 3X desired concentration "                                                                                                                                                             
#> [20] "4. Prepare H2O2 Standard Curve: 100, 50, 25, 12.5, 6.25, 3.125, 1.5625 & 0 uM in HBSS  final concentration in assay."                                                                                                 
#> [21] "Add 10 uL 30% H2O2 to 10 mL HBSS (8.82 mM)"                                                                                                                                                                           
#> [22] "Add 340 uL of 8.82 mM H2O2 to 9.660 mL HBSS (300 uM)"                                                                                                                                                                 
#> [23] "Serially dilute 1:2 the 300 uM solution in HBSS to generate 150, 75, 37.5, 18.75, 9.375, 4.6875 & 0 uM standards."                                                                                                    
#> [24] "5. Add 20 uL of HBSS to control and standard wells."                                                                                                                                                                  
#> [25] "6. Transfer 20 uL of compounds to assay plate. "                                                                                                                                                                      
#> [26] "7. Transfer 20 uL of 3 mM DTT solution in all wells containing compounds (Do not add DTT in controls or standards wells)."                                                                                            
#> [27] "8. Pre-incubate the compound and DTT at room temperature for 15 minutes."                                                                                                                                             
#> [28] "9. Meanwhile, transfer 20 uL H2O2 standards to assay plate."                                                                                                                                                          
#> [29] "Max controls: 20 uL 300 uM H2O2 in HBSS (100 uM final H2O2). "                                                                                                                                                        
#> [30] "Min controls: 20 uL of HBSS (0 uM final H2O2)."                                                                                                                                                                       
#> [31] "50% controls: 20 uL of 150 uM H2O2 in HBSS (50 uM final H2O2)."                                                                                                                                                       
#> [32] "10. Add 20 uL of 3X Phenol Red-HRP solution in HBSS. "                                                                                                                                                                
#> [33] "11. Add 10 uL of 1N NaOH stop solution and shake the plates on a shaking platform."                                                                                                                                   
#> [34] "12. Read absorbance at 610 nm on the M5 Plate Reader."                                                                                                                                                                
#> [35] " "                                                                                                                                                                                                                    
#> 
#> $`'1234'`[[1]]$assay$descr$comment
#>  [1] "List of compounds to be tested: active compounds identified in the Redox Cycling H2O2 Generation primary screen AID 878, will be confirmed in 10-point concentration response assays."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       
#>  [2] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#>  [3] "PMLSC H2O2 Assay EC50 Scoring System:"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       
#>  [4] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#>  [5] "PUBCHEM_ACTIVITY_OUTCOME"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    
#>  [6] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#>  [7] "1 - Substance is considered inactive when the EC50 is > 50 uM"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               
#>  [8] "2 - Substance is considered active when the EC50 is < 50 uM"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#>  [9] "3 - Substance activity outcome is inconclusive"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
#> [10] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [11] "PUBCHEM_ACTIVITY_SCORE"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      
#> [12] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [13] "40 - Compounds that in the presence of DTT failed to produce detectable levels of H2O2 in a 10-pt dose response assay with an EC50 > 50 uM "                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#> [14] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [15] "50 - Compounds that in the presence of DTT were active in the H2O2 generation 10-pt dose response assay with an EC50 in the 20 to 50 uM range"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               
#> [16] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [17] "60 - Compounds that in the presence of DTT were active in the H2O2 generation 10-pt dose response assay with an EC50 in the 10 to 20 uM range"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               
#> [18] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [19] "70 - Compounds that in the presence of DTT were active in the H2O2 generation 10-pt dose response assay with an EC50 in the 1 to 10 uM range"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> [20] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [21] "80 - Compounds that in the presence of DTT were active in the H2O2 generation 10-pt dose response assay with an EC50 < 1 uM"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#> [22] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [23] "Data Analysis:"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
#> [24] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [25] "Compounds will be tested in singlicate 10-point dose response assays, typically at a maximum concentration of 50uM final (1% DMSO). Compounds will be diluted to 150 uM in dH2O (3% DMSO final) and serially diluted in dH2O + 3% DMSO to provide the desired concentration range. 20uL of the compound dilutions will be transferred to a total of 60 uL assay volume (3-fold dilution)"                                                                                                                                                                                                                                                                                                                                                                    
#> [26] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [27] "An ActivityBase template has been written that will automatically calculate % activation together with plate control Signal to Noise and Z-factors for secondary dose response data. These data will be used to perform quality control (QC) and either pass or fail plates. For data from plates that pass QC, EC50#s will be calculated using the four parameter logistic model, also called the Sigmoidal Dose-Response Model. One form of the equation is: "                                                                                                                                                                                                                                                                                             
#> [28] "y = (A+((B-A)/(1+((x/C)^D))))"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               
#> [29] "where y is the percent activation and x is the corresponding concentration. The fitted C parameter is the EC50, and is defined as the concentration giving a response half way between the fitted top and bottom of the curve. The fitted D parameter is called the Slope. The A and B parameters are locked. A is assigned to be 0 and B is 100. Also reported are the R2 and Hill slope values. The R2 value is the square of the linear correlation coefficient for a given fit cell. It will return a value between 0 and 1.00 as the correlation coefficient only returns values between -1.00 and 1.00. The Hill slope value or slope value quantifies the steepness of the curve; a dose response curve with a standard slope has a Hill slope of 1.0"
#> [30] ""                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [31] "Definition of Hit Criteria: It is anticipated that all HTS actives will be confirmed in 10-pt EC50 concentration response curves. Actives confirmed as concentration dependent redox cycling H2O2 generators will be further evaluated and should exhibit: "                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#> [32] "EC50 < 50 uM"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                
#> [33] "Structural confirmation by LCMS."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            
#> [34] " "                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
#> [35] "Secondary assay testing paradigm:"                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           
#> [36] "Compounds confirmed as concentration dependent redox cycling generators of H2O2 in the presence of DTT will be tested in a concentration response assays with four alternate reducing agents; TCEP, beta-mercaptoethanol, glutathione and L-cysteine."                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       
#> 
#> $`'1234'`[[1]]$assay$descr$xref
#> $`'1234'`[[1]]$assay$descr$xref[[1]]
#> $`'1234'`[[1]]$assay$descr$xref[[1]]$xref
#>                    dburl 
#> "http://pmlsc.pitt.edu/" 
#> 
#> $`'1234'`[[1]]$assay$descr$xref[[1]]$comment
#> [1] "University of Pittsburgh Molecular Library Screening Center (PMLSC) Website"
#> 
#> 
#> $`'1234'`[[1]]$assay$descr$xref[[2]]
#> $`'1234'`[[1]]$assay$descr$xref[[2]]$xref
#>     pmid 
#> 14711389 
#> 
#> $`'1234'`[[1]]$assay$descr$xref[[2]]$comment
#> [1] "Improved statistical methods for hit selection in high-throughput screening.J Biomol Screen. 2003 Dec;8(6):634-47."
#> 
#> 
#> $`'1234'`[[1]]$assay$descr$xref[[3]]
#> $`'1234'`[[1]]$assay$descr$xref[[3]]$xref
#> aid 
#> 878 
#> 
#> $`'1234'`[[1]]$assay$descr$xref[[3]]$comment
#> [1] "NIH Compound Library Profiling: Compound and DTT Dependent Redox Cycling H2O2 Generation Primary Screen AID 878; Compounds tested at 10 uM and DTT at 0.8 mM."
#> 
#> 
#> 
#> $`'1234'`[[1]]$assay$descr$results
#> $`'1234'`[[1]]$assay$descr$results[[1]]
#> $`'1234'`[[1]]$assay$descr$results[[1]]$tid
#> [1] 1
#> 
#> $`'1234'`[[1]]$assay$descr$results[[1]]$name
#> [1] "Qualifier"
#> 
#> $`'1234'`[[1]]$assay$descr$results[[1]]$description
#> [1] "this qualifier is intended to be interpreted with the next TID called \"DR EC50 uM\". If qualifier is \"=\" than \"DR EC50 uM\" equals to the value in that column, if qualifier is \">\" than \"DR EC50 uM\" is bigger than that value"
#> 
#> $`'1234'`[[1]]$assay$descr$results[[1]]$type
#> [1] 4
#> 
#> 
#> $`'1234'`[[1]]$assay$descr$results[[2]]
#> $`'1234'`[[1]]$assay$descr$results[[2]]$tid
#> [1] 2
#> 
#> $`'1234'`[[1]]$assay$descr$results[[2]]$name
#> [1] "DR EC50 uM"
#> 
#> $`'1234'`[[1]]$assay$descr$results[[2]]$description
#> [1] "EC50 determined from a 10-point concentration response assay"
#> [2] " "                                                           
#> 
#> $`'1234'`[[1]]$assay$descr$results[[2]]$type
#> [1] 1
#> 
#> $`'1234'`[[1]]$assay$descr$results[[2]]$unit
#> [1] 5
#> 
#> $`'1234'`[[1]]$assay$descr$results[[2]]$ac
#> [1] TRUE
#> 
#> 
#> $`'1234'`[[1]]$assay$descr$results[[3]]
#> $`'1234'`[[1]]$assay$descr$results[[3]]$tid
#> [1] 3
#> 
#> $`'1234'`[[1]]$assay$descr$results[[3]]$name
#> [1] "DR EC50 R2"
#> 
#> $`'1234'`[[1]]$assay$descr$results[[3]]$description
#> [1] "R2 value, the square of the linear correlation coefficient for a given fit cell"
#> 
#> $`'1234'`[[1]]$assay$descr$results[[3]]$type
#> [1] 1
#> 
#> $`'1234'`[[1]]$assay$descr$results[[3]]$unit
#> [1] 254
#> 
#> 
#> $`'1234'`[[1]]$assay$descr$results[[4]]
#> $`'1234'`[[1]]$assay$descr$results[[4]]$tid
#> [1] 4
#> 
#> $`'1234'`[[1]]$assay$descr$results[[4]]$name
#> [1] "DR EC50 Hill Slope"
#> 
#> $`'1234'`[[1]]$assay$descr$results[[4]]$description
#> [1] "Hill slope value or slope value quantifies the steepness of the curve; a dose response curve with a standard slope has a Hill slope of 1.0 "
#> 
#> $`'1234'`[[1]]$assay$descr$results[[4]]$type
#> [1] 1
#> 
#> $`'1234'`[[1]]$assay$descr$results[[4]]$unit
#> [1] 254
#> 
#> 
#> $`'1234'`[[1]]$assay$descr$results[[5]]
#> $`'1234'`[[1]]$assay$descr$results[[5]]$tid
#> [1] 5
#> 
#> $`'1234'`[[1]]$assay$descr$results[[5]]$name
#> [1] "DR % Activation at 50 uM"
#> 
#> $`'1234'`[[1]]$assay$descr$results[[5]]$description
#> [1] "Dose response (n=1) % activation of the assay at 50 uM compound calculated from the mean max and min plate controls"
#> 
#> $`'1234'`[[1]]$assay$descr$results[[5]]$type
#> [1] 1
#> 
#> $`'1234'`[[1]]$assay$descr$results[[5]]$unit
#> [1] 15
#> 
#> $`'1234'`[[1]]$assay$descr$results[[5]]$tc
#> concentration          unit 
#>            50             5 
#> 
#> 
#> $`'1234'`[[1]]$assay$descr$results[[6]]
#> $`'1234'`[[1]]$assay$descr$results[[6]]$tid
#> [1] 6
#> 
#> $`'1234'`[[1]]$assay$descr$results[[6]]$name
#> [1] "DR Plate Mean max signal"
#> 
#> $`'1234'`[[1]]$assay$descr$results[[6]]$description
#> [1] "Dose response plate Mean maximum assay signal window plate controls n=32"
#> 
#> $`'1234'`[[1]]$assay$descr$results[[6]]$type
#> [1] 1
#> 
#> $`'1234'`[[1]]$assay$descr$results[[6]]$unit
#> [1] 254
#> 
#> 
#> $`'1234'`[[1]]$assay$descr$results[[7]]
#> $`'1234'`[[1]]$assay$descr$results[[7]]$tid
#> [1] 7
#> 
#> $`'1234'`[[1]]$assay$descr$results[[7]]$name
#> [1] "DR Plate Mean min signal"
#> 
#> $`'1234'`[[1]]$assay$descr$results[[7]]$description
#> [1] "Dose response plate Mean minimum assay signal window plate controls n=24"
#> 
#> $`'1234'`[[1]]$assay$descr$results[[7]]$type
#> [1] 1
#> 
#> $`'1234'`[[1]]$assay$descr$results[[7]]$unit
#> [1] 254
#> 
#> 
#> $`'1234'`[[1]]$assay$descr$results[[8]]
#> $`'1234'`[[1]]$assay$descr$results[[8]]$tid
#> [1] 8
#> 
#> $`'1234'`[[1]]$assay$descr$results[[8]]$name
#> [1] "DR plate Z'-factor"
#> 
#> $`'1234'`[[1]]$assay$descr$results[[8]]$description
#> [1] "Dose response plate Z'-factor calculated from the maximum and mimimum assay signal window plate controls"
#> 
#> $`'1234'`[[1]]$assay$descr$results[[8]]$type
#> [1] 1
#> 
#> $`'1234'`[[1]]$assay$descr$results[[8]]$unit
#> [1] 254
#> 
#> 
#> $`'1234'`[[1]]$assay$descr$results[[9]]
#> $`'1234'`[[1]]$assay$descr$results[[9]]$tid
#> [1] 9
#> 
#> $`'1234'`[[1]]$assay$descr$results[[9]]$name
#> [1] "DR Assay Date"
#> 
#> $`'1234'`[[1]]$assay$descr$results[[9]]$description
#> [1] "Date the dose response assay was performed"
#> 
#> $`'1234'`[[1]]$assay$descr$results[[9]]$type
#> [1] 4
#> 
#> 
#> 
#> $`'1234'`[[1]]$assay$descr$revision
#> [1] 1
#> 
#> $`'1234'`[[1]]$assay$descr$activity_outcome_method
#> [1] 2
#> 
#> $`'1234'`[[1]]$assay$descr$project_category
#> [1] 1
#> 
#> 
#> 
#> 
#>