Retrieve Compounds from PubChem

This function sends a request to PubChem to retrieve compound data based on the specified parameters. It returns a list of compounds corresponding to the provided identifiers.

Usage

get_compounds(
  identifier,
  namespace = "cid",
  operation = NULL,
  searchtype = NULL,
  options = NULL
)

Arguments

identifier

A vector of positive integers (e.g. cid, sid, aid) or identifier strings (source, inchikey, formula). In some cases, only a single identifier string (name, smiles, xref; inchi, sdf by POST only).

namespace

Specifies the namespace for the query. For the 'compound' domain, possible values include 'cid', 'name', 'smiles', 'inchi', 'sdf', 'inchikey', 'formula', 'substructure', 'superstructure', 'similarity', 'identity', 'xref', 'listkey', 'fastidentity', 'fastsimilarity_2d', 'fastsimilarity_3d', 'fastsubstructure', 'fastsuperstructure', and 'fastformula'. For other domains, the possible namespaces are domain-specific.

operation

The operation to be performed (default: NULL).

searchtype

Specifies the type of search to be performed. For structure searches, possible values are combinations of 'substructure', 'superstructure', 'similarity', 'identity' with 'smiles', 'inchi', 'sdf', 'cid'. For fast searches, possible values are combinations of 'fastidentity', 'fastsimilarity_2d', 'fastsimilarity_3d', 'fastsubstructure', 'fastsuperstructure' with 'smiles', 'smarts', 'inchi', 'sdf', 'cid', or 'fastformula'.

options

Additional parameters passed to get_json.

Value

A named list where each element corresponds to a compound retrieved from PubChem. The names of the list elements are based on the provided identifiers. If no compound is found for a given identifier, the corresponding list element will contain the string "No compound".

Examples

get_compounds(
  identifier = "aspirin",
  namespace = "name"
)
#> $`'aspirin'`
#> $`'aspirin'`[[1]]
#> $`'aspirin'`[[1]]$id
#> $`'aspirin'`[[1]]$id$id
#>  cid 
#> 2244 
#> 
#> 
#> $`'aspirin'`[[1]]$atoms
#> $`'aspirin'`[[1]]$atoms$aid
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21
#> 
#> $`'aspirin'`[[1]]$atoms$element
#>  [1] 8 8 8 8 6 6 6 6 6 6 6 6 6 1 1 1 1 1 1 1 1
#> 
#> 
#> $`'aspirin'`[[1]]$bonds
#> $`'aspirin'`[[1]]$bonds$aid1
#>  [1]  1  1  2  2  3  4  5  5  6  6  7  7  8  8  9  9 10 12 13 13 13
#> 
#> $`'aspirin'`[[1]]$bonds$aid2
#>  [1]  5 12 11 21 11 12  6  7  8 11  9 14 10 15 10 16 17 13 18 19 20
#> 
#> $`'aspirin'`[[1]]$bonds$order
#>  [1] 1 1 1 1 2 2 1 2 2 1 1 1 1 1 2 1 1 1 1 1 1
#> 
#> 
#> $`'aspirin'`[[1]]$coords
#> $`'aspirin'`[[1]]$coords[[1]]
#> $`'aspirin'`[[1]]$coords[[1]]$type
#> [1]   1   5 255
#> 
#> $`'aspirin'`[[1]]$coords[[1]]$aid
#>  [1]  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21
#> 
#> $`'aspirin'`[[1]]$coords[[1]]$conformers
#> $`'aspirin'`[[1]]$coords[[1]]$conformers[[1]]
#> $`'aspirin'`[[1]]$coords[[1]]$conformers[[1]]$x
#>  [1] 3.7320 6.3301 4.5981 2.8660 4.5981 5.4641 4.5981 6.3301 5.4641 6.3301
#> [11] 5.4641 2.8660 2.0000 4.0611 6.8671 5.4641 6.8671 2.3100 1.4631 1.6900
#> [21] 6.3301
#> 
#> $`'aspirin'`[[1]]$coords[[1]]$conformers[[1]]$y
#>  [1] -0.0600  1.4400  1.4400 -1.5600 -0.5600 -0.0600 -1.5600 -0.5600 -2.0600
#> [10] -1.5600  0.9400 -0.5600 -0.0600 -1.8700 -0.2500 -2.6800 -1.8700  0.4769
#> [19]  0.2500 -0.5969  2.0600
#> 
#> $`'aspirin'`[[1]]$coords[[1]]$conformers[[1]]$style
#> $`'aspirin'`[[1]]$coords[[1]]$conformers[[1]]$style$annotation
#> [1] 8 8 8 8 8 8
#> 
#> $`'aspirin'`[[1]]$coords[[1]]$conformers[[1]]$style$aid1
#> [1] 5 5 6 7 8 9
#> 
#> $`'aspirin'`[[1]]$coords[[1]]$conformers[[1]]$style$aid2
#> [1]  6  7  8  9 10 10
#> 
#> 
#> 
#> 
#> 
#> 
#> $`'aspirin'`[[1]]$charge
#> [1] 0
#> 
#> $`'aspirin'`[[1]]$props
#> $`'aspirin'`[[1]]$props[[1]]
#> $`'aspirin'`[[1]]$props[[1]]$urn
#> $`'aspirin'`[[1]]$props[[1]]$urn$label
#> [1] "Compound"
#> 
#> $`'aspirin'`[[1]]$props[[1]]$urn$name
#> [1] "Canonicalized"
#> 
#> $`'aspirin'`[[1]]$props[[1]]$urn$datatype
#> [1] 5
#> 
#> $`'aspirin'`[[1]]$props[[1]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[1]]$value
#> ival 
#>    1 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[2]]
#> $`'aspirin'`[[1]]$props[[2]]$urn
#> $`'aspirin'`[[1]]$props[[2]]$urn$label
#> [1] "Compound Complexity"
#> 
#> $`'aspirin'`[[1]]$props[[2]]$urn$datatype
#> [1] 7
#> 
#> $`'aspirin'`[[1]]$props[[2]]$urn$implementation
#> [1] "E_COMPLEXITY"
#> 
#> $`'aspirin'`[[1]]$props[[2]]$urn$version
#> [1] "3.4.8.18"
#> 
#> $`'aspirin'`[[1]]$props[[2]]$urn$software
#> [1] "Cactvs"
#> 
#> $`'aspirin'`[[1]]$props[[2]]$urn$source
#> [1] "Xemistry GmbH"
#> 
#> $`'aspirin'`[[1]]$props[[2]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[2]]$value
#> fval 
#>  212 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[3]]
#> $`'aspirin'`[[1]]$props[[3]]$urn
#> $`'aspirin'`[[1]]$props[[3]]$urn$label
#> [1] "Count"
#> 
#> $`'aspirin'`[[1]]$props[[3]]$urn$name
#> [1] "Hydrogen Bond Acceptor"
#> 
#> $`'aspirin'`[[1]]$props[[3]]$urn$datatype
#> [1] 5
#> 
#> $`'aspirin'`[[1]]$props[[3]]$urn$implementation
#> [1] "E_NHACCEPTORS"
#> 
#> $`'aspirin'`[[1]]$props[[3]]$urn$version
#> [1] "3.4.8.18"
#> 
#> $`'aspirin'`[[1]]$props[[3]]$urn$software
#> [1] "Cactvs"
#> 
#> $`'aspirin'`[[1]]$props[[3]]$urn$source
#> [1] "Xemistry GmbH"
#> 
#> $`'aspirin'`[[1]]$props[[3]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[3]]$value
#> ival 
#>    4 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[4]]
#> $`'aspirin'`[[1]]$props[[4]]$urn
#> $`'aspirin'`[[1]]$props[[4]]$urn$label
#> [1] "Count"
#> 
#> $`'aspirin'`[[1]]$props[[4]]$urn$name
#> [1] "Hydrogen Bond Donor"
#> 
#> $`'aspirin'`[[1]]$props[[4]]$urn$datatype
#> [1] 5
#> 
#> $`'aspirin'`[[1]]$props[[4]]$urn$implementation
#> [1] "E_NHDONORS"
#> 
#> $`'aspirin'`[[1]]$props[[4]]$urn$version
#> [1] "3.4.8.18"
#> 
#> $`'aspirin'`[[1]]$props[[4]]$urn$software
#> [1] "Cactvs"
#> 
#> $`'aspirin'`[[1]]$props[[4]]$urn$source
#> [1] "Xemistry GmbH"
#> 
#> $`'aspirin'`[[1]]$props[[4]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[4]]$value
#> ival 
#>    1 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[5]]
#> $`'aspirin'`[[1]]$props[[5]]$urn
#> $`'aspirin'`[[1]]$props[[5]]$urn$label
#> [1] "Count"
#> 
#> $`'aspirin'`[[1]]$props[[5]]$urn$name
#> [1] "Rotatable Bond"
#> 
#> $`'aspirin'`[[1]]$props[[5]]$urn$datatype
#> [1] 5
#> 
#> $`'aspirin'`[[1]]$props[[5]]$urn$implementation
#> [1] "E_NROTBONDS"
#> 
#> $`'aspirin'`[[1]]$props[[5]]$urn$version
#> [1] "3.4.8.18"
#> 
#> $`'aspirin'`[[1]]$props[[5]]$urn$software
#> [1] "Cactvs"
#> 
#> $`'aspirin'`[[1]]$props[[5]]$urn$source
#> [1] "Xemistry GmbH"
#> 
#> $`'aspirin'`[[1]]$props[[5]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[5]]$value
#> ival 
#>    3 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[6]]
#> $`'aspirin'`[[1]]$props[[6]]$urn
#> $`'aspirin'`[[1]]$props[[6]]$urn$label
#> [1] "Fingerprint"
#> 
#> $`'aspirin'`[[1]]$props[[6]]$urn$name
#> [1] "SubStructure Keys"
#> 
#> $`'aspirin'`[[1]]$props[[6]]$urn$datatype
#> [1] 16
#> 
#> $`'aspirin'`[[1]]$props[[6]]$urn$parameters
#> [1] "extended 2"
#> 
#> $`'aspirin'`[[1]]$props[[6]]$urn$implementation
#> [1] "E_SCREEN"
#> 
#> $`'aspirin'`[[1]]$props[[6]]$urn$version
#> [1] "3.4.8.18"
#> 
#> $`'aspirin'`[[1]]$props[[6]]$urn$software
#> [1] "Cactvs"
#> 
#> $`'aspirin'`[[1]]$props[[6]]$urn$source
#> [1] "Xemistry GmbH"
#> 
#> $`'aspirin'`[[1]]$props[[6]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[6]]$value
#>                                                                                                                                                                                                                                   binary 
#> "00000371C0703800000000000000000000000000000000000000300000000000000000010000001A00000800000C04809800320E80000600880220D208000208002420000888010608C80C273684351A827B60A5E01108B98788C8208E00000000000800000000000000100000000000000000" 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[7]]
#> $`'aspirin'`[[1]]$props[[7]]$urn
#> $`'aspirin'`[[1]]$props[[7]]$urn$label
#> [1] "IUPAC Name"
#> 
#> $`'aspirin'`[[1]]$props[[7]]$urn$name
#> [1] "Allowed"
#> 
#> $`'aspirin'`[[1]]$props[[7]]$urn$datatype
#> [1] 1
#> 
#> $`'aspirin'`[[1]]$props[[7]]$urn$version
#> [1] "2.7.0"
#> 
#> $`'aspirin'`[[1]]$props[[7]]$urn$software
#> [1] "Lexichem TK"
#> 
#> $`'aspirin'`[[1]]$props[[7]]$urn$source
#> [1] "OpenEye Scientific Software"
#> 
#> $`'aspirin'`[[1]]$props[[7]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[7]]$value
#>                    sval 
#> "2-acetoxybenzoic acid" 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[8]]
#> $`'aspirin'`[[1]]$props[[8]]$urn
#> $`'aspirin'`[[1]]$props[[8]]$urn$label
#> [1] "IUPAC Name"
#> 
#> $`'aspirin'`[[1]]$props[[8]]$urn$name
#> [1] "CAS-like Style"
#> 
#> $`'aspirin'`[[1]]$props[[8]]$urn$datatype
#> [1] 1
#> 
#> $`'aspirin'`[[1]]$props[[8]]$urn$version
#> [1] "2.7.0"
#> 
#> $`'aspirin'`[[1]]$props[[8]]$urn$software
#> [1] "Lexichem TK"
#> 
#> $`'aspirin'`[[1]]$props[[8]]$urn$source
#> [1] "OpenEye Scientific Software"
#> 
#> $`'aspirin'`[[1]]$props[[8]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[8]]$value
#>                      sval 
#> "2-acetyloxybenzoic acid" 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[9]]
#> $`'aspirin'`[[1]]$props[[9]]$urn
#> $`'aspirin'`[[1]]$props[[9]]$urn$label
#> [1] "IUPAC Name"
#> 
#> $`'aspirin'`[[1]]$props[[9]]$urn$name
#> [1] "Markup"
#> 
#> $`'aspirin'`[[1]]$props[[9]]$urn$datatype
#> [1] 1
#> 
#> $`'aspirin'`[[1]]$props[[9]]$urn$version
#> [1] "2.7.0"
#> 
#> $`'aspirin'`[[1]]$props[[9]]$urn$software
#> [1] "Lexichem TK"
#> 
#> $`'aspirin'`[[1]]$props[[9]]$urn$source
#> [1] "OpenEye Scientific Software"
#> 
#> $`'aspirin'`[[1]]$props[[9]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[9]]$value
#>                      sval 
#> "2-acetyloxybenzoic acid" 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[10]]
#> $`'aspirin'`[[1]]$props[[10]]$urn
#> $`'aspirin'`[[1]]$props[[10]]$urn$label
#> [1] "IUPAC Name"
#> 
#> $`'aspirin'`[[1]]$props[[10]]$urn$name
#> [1] "Preferred"
#> 
#> $`'aspirin'`[[1]]$props[[10]]$urn$datatype
#> [1] 1
#> 
#> $`'aspirin'`[[1]]$props[[10]]$urn$version
#> [1] "2.7.0"
#> 
#> $`'aspirin'`[[1]]$props[[10]]$urn$software
#> [1] "Lexichem TK"
#> 
#> $`'aspirin'`[[1]]$props[[10]]$urn$source
#> [1] "OpenEye Scientific Software"
#> 
#> $`'aspirin'`[[1]]$props[[10]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[10]]$value
#>                      sval 
#> "2-acetyloxybenzoic acid" 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[11]]
#> $`'aspirin'`[[1]]$props[[11]]$urn
#> $`'aspirin'`[[1]]$props[[11]]$urn$label
#> [1] "IUPAC Name"
#> 
#> $`'aspirin'`[[1]]$props[[11]]$urn$name
#> [1] "Systematic"
#> 
#> $`'aspirin'`[[1]]$props[[11]]$urn$datatype
#> [1] 1
#> 
#> $`'aspirin'`[[1]]$props[[11]]$urn$version
#> [1] "2.7.0"
#> 
#> $`'aspirin'`[[1]]$props[[11]]$urn$software
#> [1] "Lexichem TK"
#> 
#> $`'aspirin'`[[1]]$props[[11]]$urn$source
#> [1] "OpenEye Scientific Software"
#> 
#> $`'aspirin'`[[1]]$props[[11]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[11]]$value
#>                      sval 
#> "2-acetyloxybenzoic acid" 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[12]]
#> $`'aspirin'`[[1]]$props[[12]]$urn
#> $`'aspirin'`[[1]]$props[[12]]$urn$label
#> [1] "IUPAC Name"
#> 
#> $`'aspirin'`[[1]]$props[[12]]$urn$name
#> [1] "Traditional"
#> 
#> $`'aspirin'`[[1]]$props[[12]]$urn$datatype
#> [1] 1
#> 
#> $`'aspirin'`[[1]]$props[[12]]$urn$version
#> [1] "2.7.0"
#> 
#> $`'aspirin'`[[1]]$props[[12]]$urn$software
#> [1] "Lexichem TK"
#> 
#> $`'aspirin'`[[1]]$props[[12]]$urn$source
#> [1] "OpenEye Scientific Software"
#> 
#> $`'aspirin'`[[1]]$props[[12]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[12]]$value
#>                    sval 
#> "2-acetoxybenzoic acid" 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[13]]
#> $`'aspirin'`[[1]]$props[[13]]$urn
#> $`'aspirin'`[[1]]$props[[13]]$urn$label
#> [1] "InChI"
#> 
#> $`'aspirin'`[[1]]$props[[13]]$urn$name
#> [1] "Standard"
#> 
#> $`'aspirin'`[[1]]$props[[13]]$urn$datatype
#> [1] 1
#> 
#> $`'aspirin'`[[1]]$props[[13]]$urn$version
#> [1] "1.0.6"
#> 
#> $`'aspirin'`[[1]]$props[[13]]$urn$software
#> [1] "InChI"
#> 
#> $`'aspirin'`[[1]]$props[[13]]$urn$source
#> [1] "iupac.org"
#> 
#> $`'aspirin'`[[1]]$props[[13]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[13]]$value
#>                                                                   sval 
#> "InChI=1S/C9H8O4/c1-6(10)13-8-5-3-2-4-7(8)9(11)12/h2-5H,1H3,(H,11,12)" 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[14]]
#> $`'aspirin'`[[1]]$props[[14]]$urn
#> $`'aspirin'`[[1]]$props[[14]]$urn$label
#> [1] "InChIKey"
#> 
#> $`'aspirin'`[[1]]$props[[14]]$urn$name
#> [1] "Standard"
#> 
#> $`'aspirin'`[[1]]$props[[14]]$urn$datatype
#> [1] 1
#> 
#> $`'aspirin'`[[1]]$props[[14]]$urn$version
#> [1] "1.0.6"
#> 
#> $`'aspirin'`[[1]]$props[[14]]$urn$software
#> [1] "InChI"
#> 
#> $`'aspirin'`[[1]]$props[[14]]$urn$source
#> [1] "iupac.org"
#> 
#> $`'aspirin'`[[1]]$props[[14]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[14]]$value
#>                          sval 
#> "BSYNRYMUTXBXSQ-UHFFFAOYSA-N" 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[15]]
#> $`'aspirin'`[[1]]$props[[15]]$urn
#> $`'aspirin'`[[1]]$props[[15]]$urn$label
#> [1] "Log P"
#> 
#> $`'aspirin'`[[1]]$props[[15]]$urn$name
#> [1] "XLogP3"
#> 
#> $`'aspirin'`[[1]]$props[[15]]$urn$datatype
#> [1] 7
#> 
#> $`'aspirin'`[[1]]$props[[15]]$urn$version
#> [1] "3.0"
#> 
#> $`'aspirin'`[[1]]$props[[15]]$urn$source
#> [1] "sioc-ccbg.ac.cn"
#> 
#> $`'aspirin'`[[1]]$props[[15]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[15]]$value
#> fval 
#>  1.2 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[16]]
#> $`'aspirin'`[[1]]$props[[16]]$urn
#> $`'aspirin'`[[1]]$props[[16]]$urn$label
#> [1] "Mass"
#> 
#> $`'aspirin'`[[1]]$props[[16]]$urn$name
#> [1] "Exact"
#> 
#> $`'aspirin'`[[1]]$props[[16]]$urn$datatype
#> [1] 1
#> 
#> $`'aspirin'`[[1]]$props[[16]]$urn$version
#> [1] "2.2"
#> 
#> $`'aspirin'`[[1]]$props[[16]]$urn$software
#> [1] "PubChem"
#> 
#> $`'aspirin'`[[1]]$props[[16]]$urn$source
#> [1] "ncbi.nlm.nih.gov"
#> 
#> $`'aspirin'`[[1]]$props[[16]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[16]]$value
#>           sval 
#> "180.04225873" 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[17]]
#> $`'aspirin'`[[1]]$props[[17]]$urn
#> $`'aspirin'`[[1]]$props[[17]]$urn$label
#> [1] "Molecular Formula"
#> 
#> $`'aspirin'`[[1]]$props[[17]]$urn$datatype
#> [1] 1
#> 
#> $`'aspirin'`[[1]]$props[[17]]$urn$version
#> [1] "2.2"
#> 
#> $`'aspirin'`[[1]]$props[[17]]$urn$software
#> [1] "PubChem"
#> 
#> $`'aspirin'`[[1]]$props[[17]]$urn$source
#> [1] "ncbi.nlm.nih.gov"
#> 
#> $`'aspirin'`[[1]]$props[[17]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[17]]$value
#>     sval 
#> "C9H8O4" 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[18]]
#> $`'aspirin'`[[1]]$props[[18]]$urn
#> $`'aspirin'`[[1]]$props[[18]]$urn$label
#> [1] "Molecular Weight"
#> 
#> $`'aspirin'`[[1]]$props[[18]]$urn$datatype
#> [1] 1
#> 
#> $`'aspirin'`[[1]]$props[[18]]$urn$version
#> [1] "2.2"
#> 
#> $`'aspirin'`[[1]]$props[[18]]$urn$software
#> [1] "PubChem"
#> 
#> $`'aspirin'`[[1]]$props[[18]]$urn$source
#> [1] "ncbi.nlm.nih.gov"
#> 
#> $`'aspirin'`[[1]]$props[[18]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[18]]$value
#>     sval 
#> "180.16" 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[19]]
#> $`'aspirin'`[[1]]$props[[19]]$urn
#> $`'aspirin'`[[1]]$props[[19]]$urn$label
#> [1] "SMILES"
#> 
#> $`'aspirin'`[[1]]$props[[19]]$urn$name
#> [1] "Canonical"
#> 
#> $`'aspirin'`[[1]]$props[[19]]$urn$datatype
#> [1] 1
#> 
#> $`'aspirin'`[[1]]$props[[19]]$urn$version
#> [1] "2.3.0"
#> 
#> $`'aspirin'`[[1]]$props[[19]]$urn$software
#> [1] "OEChem"
#> 
#> $`'aspirin'`[[1]]$props[[19]]$urn$source
#> [1] "OpenEye Scientific Software"
#> 
#> $`'aspirin'`[[1]]$props[[19]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[19]]$value
#>                       sval 
#> "CC(=O)OC1=CC=CC=C1C(=O)O" 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[20]]
#> $`'aspirin'`[[1]]$props[[20]]$urn
#> $`'aspirin'`[[1]]$props[[20]]$urn$label
#> [1] "SMILES"
#> 
#> $`'aspirin'`[[1]]$props[[20]]$urn$name
#> [1] "Isomeric"
#> 
#> $`'aspirin'`[[1]]$props[[20]]$urn$datatype
#> [1] 1
#> 
#> $`'aspirin'`[[1]]$props[[20]]$urn$version
#> [1] "2.3.0"
#> 
#> $`'aspirin'`[[1]]$props[[20]]$urn$software
#> [1] "OEChem"
#> 
#> $`'aspirin'`[[1]]$props[[20]]$urn$source
#> [1] "OpenEye Scientific Software"
#> 
#> $`'aspirin'`[[1]]$props[[20]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[20]]$value
#>                       sval 
#> "CC(=O)OC1=CC=CC=C1C(=O)O" 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[21]]
#> $`'aspirin'`[[1]]$props[[21]]$urn
#> $`'aspirin'`[[1]]$props[[21]]$urn$label
#> [1] "Topological"
#> 
#> $`'aspirin'`[[1]]$props[[21]]$urn$name
#> [1] "Polar Surface Area"
#> 
#> $`'aspirin'`[[1]]$props[[21]]$urn$datatype
#> [1] 7
#> 
#> $`'aspirin'`[[1]]$props[[21]]$urn$implementation
#> [1] "E_TPSA"
#> 
#> $`'aspirin'`[[1]]$props[[21]]$urn$version
#> [1] "3.4.8.18"
#> 
#> $`'aspirin'`[[1]]$props[[21]]$urn$software
#> [1] "Cactvs"
#> 
#> $`'aspirin'`[[1]]$props[[21]]$urn$source
#> [1] "Xemistry GmbH"
#> 
#> $`'aspirin'`[[1]]$props[[21]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[21]]$value
#> fval 
#> 63.6 
#> 
#> 
#> $`'aspirin'`[[1]]$props[[22]]
#> $`'aspirin'`[[1]]$props[[22]]$urn
#> $`'aspirin'`[[1]]$props[[22]]$urn$label
#> [1] "Weight"
#> 
#> $`'aspirin'`[[1]]$props[[22]]$urn$name
#> [1] "MonoIsotopic"
#> 
#> $`'aspirin'`[[1]]$props[[22]]$urn$datatype
#> [1] 1
#> 
#> $`'aspirin'`[[1]]$props[[22]]$urn$version
#> [1] "2.2"
#> 
#> $`'aspirin'`[[1]]$props[[22]]$urn$software
#> [1] "PubChem"
#> 
#> $`'aspirin'`[[1]]$props[[22]]$urn$source
#> [1] "ncbi.nlm.nih.gov"
#> 
#> $`'aspirin'`[[1]]$props[[22]]$urn$release
#> [1] "2021.10.14"
#> 
#> 
#> $`'aspirin'`[[1]]$props[[22]]$value
#>           sval 
#> "180.04225873" 
#> 
#> 
#> 
#> $`'aspirin'`[[1]]$count
#>        heavy_atom       atom_chiral   atom_chiral_def atom_chiral_undef 
#>                13                 0                 0                 0 
#>       bond_chiral   bond_chiral_def bond_chiral_undef      isotope_atom 
#>                 0                 0                 0                 0 
#>     covalent_unit         tautomers 
#>                 1                -1 
#> 
#> 
#>